code review: magic numbers, ThemeProvider, class for accessToken, LayoutAuth, devide types
