+++ src/components/LayoutAuth/LayoutAuth.module.scss
@@ -1,0 +1,31 @@
+.auth {
+  font-family: Nunito Sans;
+  width: 100%;
+  height: 100vh;
+
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+
+  .imageWrapper {
+    height: 100vh;
+    width: 300px;
+    width: 1109px;
+
+
+    position: relative;
+    overflow: hidden;
+    // padding-bottom: 101.142857%;
+    // 98,8700565
+
+    .image {
+      position: absolute;
+      top: 0;
+      left: 0;
+      width: 100%;
+      height: 100%;
+      object-fit: cover;
+    }
+  }
+
+}
