+++ src/pages/RegisterPage/RegisterPage.module.scss
@@ -1,128 +1,95 @@
 .register {
-  font-family: Nunito Sans;
-  width: 100%;
-  height: 100vh;
-
   display: flex;
-  flex-direction: row;
-  justify-content: space-between;
+  flex-direction: column;
+  height: 100vh;
+  width: 811px;
+  justify-content: center;
+  align-items: center;
 
-  .imageWrapper {
-    height: 100vh;
-    width: 300px;
-    width: 1109px;
-
-
-    position: relative;
-    overflow: hidden;
-    // padding-bottom: 101.142857%;
-    // 98,8700565
-
-    .image {
-      position: absolute;
-      top: 0;
-      left: 0;
-      width: 100%;
-      height: 100%;
-      object-fit: cover;
-    }
-  }
-
-  .cardWrapper {
+  .card {
+    width: 420px;
     display: flex;
     flex-direction: column;
-    height: 100vh;
-    width: 811px;
-    justify-content: center;
     align-items: center;
+    // gap: 36px;
+    gap: 20px;
 
-    .card {
-      width: 420px;
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-      // gap: 36px;
-      gap: 20px;
 
   
-    
-  
-      .header {
-        width: 100%;
-  
-        .h1 {
-          font-weight: 700;
-          font-size: 36px;
-          color: rgba(82, 82, 82, 1);
-          margin-block-start: 0;
-          margin-block-end: 0;
-          margin-left: 26.5px;
-          margin-right: 26.5px;
-          // margin-top: 36px;
-          line-height: 1.3;
-        }
-        .p {
-          font-weight: 400;
-          font-size: 16px;
-          margin-block-start: 0;
-          margin-block-end: 0;
-          margin-left: 26.5px;
-          margin-right: 26.5px;
-        }
-  
+
+    .header {
+      width: 100%;
+
+      .h1 {
+        font-weight: 700;
+        font-size: 36px;
+        color: rgba(82, 82, 82, 1);
+        margin-block-start: 0;
+        margin-block-end: 0;
+        margin-left: 26.5px;
+        margin-right: 26.5px;
+        // margin-top: 36px;
+        line-height: 1.3;
       }
-  
-      .registerForm {
-        width: 100%;
-  
-        .customInput {}
-        .customButton {
-          display: flex;
-          flex-direction: row;
-          justify-content: center;
-          align-items: center;
-          width: 100%;
-          height: 50px;
-        }
-  
-        :global(.ant-form-item) {
-          margin-bottom: 7px;
-        }
+      .p {
+        font-weight: 400;
+        font-size: 16px;
+        margin-block-start: 0;
+        margin-block-end: 0;
+        margin-left: 26.5px;
+        margin-right: 26.5px;
       }
 
-      .registerLink {
+    }
+
+    .registerForm {
+      width: 100%;
+
+      .customInput {}
+      .customButton {
         display: flex;
         flex-direction: row;
         justify-content: center;
-        gap: 5px;
-  
-        .text {
-          font-family: Nunito Sans;
-          font-weight: 400;
-          font-size: 18px;
-          line-height: 100%;
-          letter-spacing: 0%;
-          color: rgba(130, 130, 130, 1);
-          margin-block-start: 0;
-          margin-block-end: 0;
-        }
-  
-        .link {
-          display: flex;
-          flex-direction: row;
-          align-items: center;
-  
-          font-family: Nunito Sans;
-          font-weight: 600;
-          font-size: 18px;
-          line-height: 100%;
-          letter-spacing: 0%;
-          color: rgba(127, 38, 91, 1);
-        }
+        align-items: center;
+        width: 100%;
+        height: 50px;
       }
-  
+
+      :global(.ant-form-item) {
+        margin-bottom: 7px;
+      }
     }
+
+    .registerLink {
+      display: flex;
+      flex-direction: row;
+      justify-content: center;
+      gap: 5px;
+
+      .text {
+        font-family: Nunito Sans;
+        font-weight: 400;
+        font-size: 18px;
+        line-height: 100%;
+        letter-spacing: 0%;
+        color: rgba(130, 130, 130, 1);
+        margin-block-start: 0;
+        margin-block-end: 0;
+      }
+
+      .link {
+        display: flex;
+        flex-direction: row;
+        align-items: center;
+
+        font-family: Nunito Sans;
+        font-weight: 600;
+        font-size: 18px;
+        line-height: 100%;
+        letter-spacing: 0%;
+        color: rgba(127, 38, 91, 1);
+      }
+    }
+
   }
-
-
 }
