+++ src/pages/RegisterPage/RegisterPage.tsx
@@ -6,7 +6,6 @@ import { PhoneInput } from '@/components/ui/phoneInput';
 import { openNotification } from '@/utils/errors';
 import { registerUser } from '@/api/api';
 import s from './RegisterPage.module.scss';
-import skelet from '../../assets/images/skelet.png';
 
 
 
@@ -14,6 +13,14 @@ const RegisterPage = () => {
   const [isOpenLink, setIsOpenLink] = useState<boolean>(false);
   const { Item } = Form;
   const { Password } = Input;
+
+  const MIN_SIZE_NAME = 1;
+  const MAX_SIZE_NAME = 60;
+  const MIN_SIZE_LOGIN = 2;
+  const MAX_SIZE_LOGIN = 60;
+  const MIN_SIZE_PASSWORD = 6;
+  const MAX_SIZE_PASSWORD = 60;
+
 
   type FieldType = {
     username: string;
@@ -61,138 +68,127 @@ const RegisterPage = () => {
 
   return (
     <div className={s.register}>
-      <div className={s.imageWrapper}>
-        <img
-          className={s.image}
-          src={skelet}
-          alt='skelet'
-         />
-      </div>
-      <div className={s.cardWrapper}>
-        <div className={s.card}>
-          <header className={s.header}>
-            {/* <AuthIcon /> */}
-            <h1 className={s.h1}>Register to your Account</h1>
-            <p className={s.p}>See what is going on with your business</p>
-          </header>
+      <div className={s.card}>
+        <header className={s.header}>
+          {/* <AuthIcon /> */}
+          <h1 className={s.h1}>Register to your Account</h1>
+          <p className={s.p}>See what is going on with your business</p>
+        </header>
 
-          <Form
-            className={s.registerForm}
-            name="register"
-            onFinish={onFinish}
-            onFinishFailed={onFinishFailed}
+        <Form
+          className={s.registerForm}
+          name="register"
+          onFinish={onFinish}
+          onFinishFailed={onFinishFailed}
+        >
+          <Item<FieldType>
+            label="Имя пользователя"
+            name="username"
+            layout="vertical"
+            rules={[
+              { required: true, message: 'Обязательное поле' },
+              { min: MIN_SIZE_NAME, message: `Cимволов не может быть менее ${MIN_SIZE_NAME}` },
+              { max: MAX_SIZE_NAME, message: `Cимволов не может быть более ${MAX_SIZE_NAME}` },
+              { 
+                pattern: /^[a-zA-Zа-яА-ЯёЁ ]+$/, 
+                message: 'Имя может содержать только символы русского/латинского алфавита'
+              }
+            ]}
+            style={{ flex: 1 }}
           >
-            <Item<FieldType>
-              label="Имя пользователя"
-              name="username"
+            <Input className={s.customInput} />
+          </Item>
+          <Item<FieldType>
+            label="Логин"
+            name="login"
+            layout="vertical"
+            rules={[
+              { required: true, message: 'Обязательное поле' },
+              { min: MIN_SIZE_LOGIN, message: `Cимволов не может быть менее ${MIN_SIZE_LOGIN}` },
+              { max: MAX_SIZE_LOGIN, message: `Cимволов не может быть более ${MAX_SIZE_LOGIN}` },
+              { 
+                pattern: /^[a-zA-Z ]+$/, 
+                message: 'Логин может содержать только символы латинского алфавита'
+              }
+            ]}
+            style={{ flex: 1 }}
+          >
+            <Input className={s.customInput} />
+          </Item>
+          <Item<FieldType>
+            label="Пароль"
+            name="password"
+            layout="vertical"
+            rules={[
+              { required: true, message: 'Обязательное поле' },
+              { min: MIN_SIZE_PASSWORD, message: `Cимволов не может быть менее ${MIN_SIZE_PASSWORD}` },
+              { max: MAX_SIZE_PASSWORD, message: `Cимволов не может быть более ${MAX_SIZE_PASSWORD}` },
+            ]}
+            style={{ flex: 1 }}
+          >
+            <Password className={s.customInput} />
+          </Item>
+          <Item<FieldType>
+            label="Повторите пароль"
+            name="password2"
+            layout="vertical"
+            dependencies={['password']}
+            rules={[
+              { required: true, message: 'Поле не может быть пустым!!' },
+              ({ getFieldValue }) => ({
+                validator(_, value) {
+                  if (!value || getFieldValue('password') === value) {
+                    return Promise.resolve();
+                  }
+                  return Promise.reject(new Error('Пароли не совпадают!'));
+                },
+              }),
+            ]}
+            style={{ flex: 1 }}
+          >
+            <Password className={s.customInput} />
+          </Item>
+          <Item<FieldType>
+            label="Почтовый адрес"
+            name="email"
+            layout="vertical"
+            rules={[
+              { required: true, message: 'Обязательное поле' },
+              { 
+                pattern: /^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/, 
+                message: 'Невалидный email'
+              }
+            ]}
+            style={{ flex: 1 }}
+          >
+            <Input className={s.customInput} />
+          </Item>
+          <Item<FieldType>
               layout="vertical"
-              rules={[
-                { required: true, message: 'Обязательное поле' },
-                { min: 1, message: 'Cимволов не может быть менее 1' },
-                { max: 60, message: 'Cимволов не может быть более 60' },
-                { 
-                  pattern: /^[a-zA-Zа-яА-ЯёЁ ]+$/, 
-                  message: 'Имя может содержать только символы русского/латинского алфавита'
-                }
-              ]}
-              style={{ flex: 1 }}
+              label="Телефон"
+              name="phone"
+              style={{flex: 1}}
             >
-              <Input className={s.customInput} />
+            <PhoneInput />
             </Item>
-            <Item<FieldType>
-              label="Логин"
-              name="login"
-              layout="vertical"
-              rules={[
-                { required: true, message: 'Обязательное поле' },
-                { min: 2, message: 'Cимволов не может быть менее 2' },
-                { max: 60, message: 'Cимволов не может быть более 60' },
-                { 
-                  pattern: /^[a-zA-Z ]+$/, 
-                  message: 'Логин может содержать только символы латинского алфавита'
-                }
-              ]}
-              style={{ flex: 1 }}
-            >
-              <Input className={s.customInput} />
-            </Item>
-            <Item<FieldType>
-              label="Пароль"
-              name="password"
-              layout="vertical"
-              rules={[
-                { required: true, message: 'Обязательное поле' },
-                { min: 6, message: 'Cимволов не может быть менее 6' },
-                { max: 60, message: 'Cимволов не может быть более 60' },
-              ]}
-              style={{ flex: 1 }}
-            >
-              <Password className={s.customInput} />
-            </Item>
-            <Item<FieldType>
-              label="Повторите пароль"
-              name="password2"
-              layout="vertical"
-              dependencies={['password']}
-              rules={[
-                { required: true, message: 'Поле не может быть пустым!!' },
-                ({ getFieldValue }) => ({
-                  validator(_, value) {
-                    if (!value || getFieldValue('password') === value) {
-                      return Promise.resolve();
-                    }
-                    return Promise.reject(new Error('Пароли не совпадают!'));
-                  },
-                }),
-              ]}
-              style={{ flex: 1 }}
-            >
-              <Password className={s.customInput} />
-            </Item>
-            <Item<FieldType>
-              label="Почтовый адрес"
-              name="email"
-              layout="vertical"
-              rules={[
-                { required: true, message: 'Обязательное поле' },
-                { 
-                  pattern: /^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/, 
-                  message: 'Невалидный email'
-                }
-              ]}
-              style={{ flex: 1 }}
-            >
-              <Input className={s.customInput} />
-            </Item>
-            <Item<FieldType>
-                layout="vertical"
-                label="Телефон"
-                name="phone"
-                style={{flex: 1}}
-              >
-              <PhoneInput />
-              </Item>
-            <Button 
-              type="primary" 
-              className={s.customButton} 
-              htmlType="submit"
-            >
-              Register
-            </Button>
-          </Form>
-          <div className={s.registerLink}>
-            <p className={s.text}>Already have an account?</p>
-            <Link to='/login' className={s.link}>
-              Log in
-            </Link>
-          </div>
-          { isOpenLink 
-            ? <Link to={"/login"}>
-                Перейти на страницу авторизации для входа в систему
-              </Link>
-            : <></>}
-
+          <Button 
+            type="primary" 
+            className={s.customButton} 
+            htmlType="submit"
+          >
+            Register
+          </Button>
+        </Form>
+        <div className={s.registerLink}>
+          <p className={s.text}>Already have an account?</p>
+          <Link to='/login' className={s.link}>
+            Log in
+          </Link>
         </div>
+        { isOpenLink && <Link to={"/login"}>
+            Перейти на страницу авторизации для входа в систему
+          </Link> }
+
       </div>
     </div>
   )
