+++ src/types/user.types.ts
@@ -1,34 +1,3 @@
-export interface Todo {
-	id: number;
-	title: string;
-	created: string;
-	isDone: boolean; 
-}
-
-export interface TodoRequest {
-	title?: string;
- 	isDone?: boolean;  
-}
-
-export interface TodoInfo { 
-	all: number;
-	completed: number;
-	inWork: number;
-}
-
-export type Filter = "all" | "completed" | "inWork";
-
-
-export interface MetaResponse<T, N> {
-	data: T[];
-	info?: N;
-	meta: {
-		totalAmount: number;
-	}
-}
-
-//---
-
 export interface UserRegistration { 
   login: string; 
   username: string; 
@@ -73,6 +42,3 @@ export interface Token {
 // interface PasswordRequest { 
 //   password: string; 
 // }
-
-
-
